+++ server-cpp-langgraph/client/client.cpp
@@ -10,7 +10,7 @@ using namespace std;
 
 int main(int argn, char **argv) {
 #if IS_LINUX_D
-  agentxx::logxx::signalError(argv[0]);
+  agentxx::util::signalError(argv[0]);
 #endif
   std::cout << "======= Agentxx Client =======" << std::endl;