+++ server-cpp-langgraph/lib/agent/deepagent.h
@@ -7,7 +7,7 @@
 #include "neograph/llm/openai_provider.h"
 #include "neograph/mcp/client.h"
 #include "neograph/neograph.h"
-#include "nodes/Toolcall.h"
+#include "nodes/toolcall.h"
 #include "tools/execute_command.h"
 #include "tools/filesystem.h"
 #include "tools/get_current_system_datetime.h"
@@ -45,6 +45,8 @@ public:
     tools.push_back(std::make_unique<agentxx::tools::FilesystemEditFileTool>());
     tools.push_back(std::make_unique<agentxx::tools::FilesystemGlobTool>());
     tools.push_back(std::make_unique<agentxx::tools::ExecuteCommandTool>());
+    tools.push_back(
+        std::make_unique<agentxx::tools::StringHtml2MarkdownTool>());
 
     for (auto &url : config->mcpServerUrls) {
       auto mcp_client = neograph::mcp::MCPClient{url};
@@ -77,7 +79,7 @@ public:
     //   __start__ -> llm -> (has_tool_calls ? tools : __end__)
     //                         tools -> llm  (loop back)
     auto definition = neograph::json{
-        {"name", "react_agent"},
+        {"name", config->agentName},
         {"channels", {{"messages", {{"type", "list"}, {"reducer", "append"}}}}},
         {
             "nodes",